Allow running different analysis' without need to change hardcoded lines
